APIs/presalytics.io/ooxml/0.1.0/python-aiohttp-client/tests/test_shared_line_end_sizes_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.problem_details import ProblemDetails
from openapi_server.models.shared_line_end_sizes import SharedLineEndSizes


pytestmark = pytest.mark.asyncio

async def test_shared_lineendsizes_get(client):
    """Test case for shared_lineendsizes_get

    LineEndSizes: List All Possible Types
    """
    headers = { 
        'Accept': 'application/json',
    }
    response = await client.request(
        method='GET',
        path='/ooxml-automation/Shared/LineEndSizes',
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_shared_lineendsizes_get_id(client):
    """Test case for shared_lineendsizes_get_id

    LineEndSizes: Get by Id
    """
    headers = { 
        'Accept': 'application/json',
    }
    response = await client.request(
        method='GET',
        path='/ooxml-automation/Shared/LineEndSizes/{id}'.format(id='id_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_shared_lineendsizes_typeid_get_type_id(client):
    """Test case for shared_lineendsizes_typeid_get_type_id

    LineEndSizes: Get By Type Id
    """
    headers = { 
        'Accept': 'application/json',
    }
    response = await client.request(
        method='GET',
        path='/ooxml-automation/Shared/LineEndSizes/TypeId/{type_id}'.format(type_id=56),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

